# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * University Placement Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  company: 'company',
  student: 'student',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  cgpa?: number | null;
  /** @nullable */
  graduationYear?: number | null;
}

export interface AuthResponse {
  user: User;
}

export type PostingStatus = typeof PostingStatus[keyof typeof PostingStatus];


export const PostingStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  closed: 'closed',
} as const;

export interface Posting {
  id: number;
  companyId: number;
  companyName: string;
  title: string;
  description: string;
  /** @nullable */
  eligibility?: string | null;
  /** @nullable */
  minCgpa?: number | null;
  /** @nullable */
  branches?: string | null;
  /** @nullable */
  ctc?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  slots?: number | null;
  status: PostingStatus;
  /** @nullable */
  rejectionReason?: string | null;
  deadline: string;
  createdAt: string;
  updatedAt?: string;
  /** @nullable */
  applicationCount?: number | null;
  /** @nullable */
  hasApplied?: boolean | null;
}

export type PostingDetailStatus = typeof PostingDetailStatus[keyof typeof PostingDetailStatus];


export const PostingDetailStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  closed: 'closed',
} as const;

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  applied: 'applied',
  shortlisted: 'shortlisted',
  selected: 'selected',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  postingId: number;
  /** @nullable */
  postingTitle?: string | null;
  /** @nullable */
  companyName?: string | null;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  studentEmail?: string | null;
  /** @nullable */
  studentDepartment?: string | null;
  /** @nullable */
  studentCgpa?: number | null;
  /** @nullable */
  coverLetter?: string | null;
  status: ApplicationStatus;
  appliedAt: string;
  updatedAt?: string;
}

export interface PostingDetail {
  id: number;
  companyId: number;
  companyName: string;
  title: string;
  description: string;
  /** @nullable */
  eligibility?: string | null;
  /** @nullable */
  minCgpa?: number | null;
  /** @nullable */
  branches?: string | null;
  /** @nullable */
  ctc?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  slots?: number | null;
  status: PostingDetailStatus;
  /** @nullable */
  rejectionReason?: string | null;
  deadline: string;
  createdAt: string;
  updatedAt?: string;
  applications: Application[];
}

export interface PostingInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  eligibility?: string;
  minCgpa?: number;
  branches?: string;
  ctc?: string;
  location?: string;
  slots?: number;
  deadline: string;
}

export interface PostingUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  eligibility?: string;
  minCgpa?: number;
  branches?: string;
  ctc?: string;
  location?: string;
  slots?: number;
  deadline?: string;
}

export interface RejectionInput {
  /** @minLength 1 */
  reason: string;
}

export interface ApplicationInput {
  coverLetter?: string;
}

export type ApplicationStatusUpdateStatus = typeof ApplicationStatusUpdateStatus[keyof typeof ApplicationStatusUpdateStatus];


export const ApplicationStatusUpdateStatus = {
  applied: 'applied',
  shortlisted: 'shortlisted',
  selected: 'selected',
  rejected: 'rejected',
} as const;

export interface ApplicationStatusUpdate {
  status: ApplicationStatusUpdateStatus;
}

export interface AnalyticsOverview {
  totalStudents: number;
  totalCompanies: number;
  totalPostings: number;
  approvedPostings: number;
  totalApplications: number;
  placedStudents: number;
  placementRate: number;
  pendingApprovals: number;
}

export interface CompanyAnalytics {
  companyId: number;
  companyName: string;
  postingCount: number;
  applicationCount: number;
  selectedCount: number;
}

export interface TimelinePoint {
  month: string;
  applications: number;
  placements: number;
}

export interface PostingsSummary {
  pending: number;
  approved: number;
  rejected: number;
  closed: number;
}

export interface Notification {
  id: number;
  userId: number;
  message: string;
  /** @nullable */
  type?: string | null;
  read: boolean;
  createdAt: string;
}

export type ListPostingsParams = {
status?: ListPostingsStatus;
/**
 * @nullable
 */
companyId?: number | null;
};

export type ListPostingsStatus = typeof ListPostingsStatus[keyof typeof ListPostingsStatus];


export const ListPostingsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  closed: 'closed',
} as const;