# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: University Placement Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with role-based credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /postings/{id}/applications:
    get:
      operationId: listPostingApplications
      tags: [applications]
      summary: List applications for a posting (admin or company)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications for the posting
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: applyToPosting
      tags: [applications]
      summary: Apply to a posting (student only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "409":
          description: Already applied
...
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status (company shortlists/selects/rejects)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [company, student, admin]
        companyName:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        cgpa:
          type: ["number", "null"]
        graduationYear:
          type: ["integer", "null"]
...
    Application:
      type: object
      required: [id, postingId, studentId, status, appliedAt]
      properties:
        id:
          type: integer
        postingId:
          type: integer
        postingTitle:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        studentEmail:
          type: ["string", "null"]
        studentDepartment:
          type: ["string", "null"]
        studentCgpa:
          type: ["number", "null"]
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected]
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicationInput:
      type: object
      properties:
        coverLetter:
          type: string
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected]
...