# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * University Placement Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  cgpa?: number | null;
  /** @nullable */
  graduationYear?: number | null;
}